// apps/git-client/src/generated/HostingChangeRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type HostingChangeRequest = { number: number, title: string, state: string, author: string, sourceBranch: string, targetBranch: string, webUrl: string, nodeId: string | null, draft: boolean, updatedAt: string, };